added add recurring payment to payment blocker
