+++ src/mobility/components/onboarding/PaymentScreenComponent.tsx
@@ -2,25 +2,22 @@ import React from 'react';
 import {OnboardingScreenComponent} from '@atb/onboarding';
 import {MobilityTexts} from '@atb/translations/screens/subscreens/MobilityTexts';
 import {useTranslation} from '@atb/translations';
+import {useRecurringPayment} from '@atb/ticketing/use-recurring-payment';
 import {ThemedMyLocation} from '@atb/theme/ThemedAssets';
 
 export type PaymentScreenComponentProps = {};
 
 export const PaymentScreenComponent = ({}: PaymentScreenComponentProps) => {
   const {t} = useTranslation();
+  const {onAddRecurringPayment} = useRecurringPayment();
 
   return (
     <OnboardingScreenComponent
-      illustration={
-        //Updating in another issue
-        <ThemedMyLocation height={220} />
-      }
+      illustration={<ThemedMyLocation fontSize={220} />}
       title={t(MobilityTexts.shmoRequirements.payment.title)}
       description={t(MobilityTexts.shmoRequirements.payment.description)}
       buttonText={t(MobilityTexts.shmoRequirements.payment.button)}
-      buttonOnPress={() => {
-        //coming in another issue
-      }}
+      buttonOnPress={onAddRecurringPayment}
       headerProps={{
         rightButton: {type: 'close', withIcon: true},
       }}
