+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_ProfileStack/Profile_PaymentMethodsScreen.tsx
@@ -10,102 +10,31 @@ import {
 import {ThemeText} from '@atb/components/text';
 import {ThemeIcon} from '@atb/components/theme-icon';
 import {StyleSheet, Theme, useThemeContext} from '@atb/theme';
-import {
-  addPaymentMethod,
-  humanizePaymentType,
-  RecurringPayment,
-} from '@atb/ticketing';
+import {humanizePaymentType, RecurringPayment} from '@atb/ticketing';
 import {useTranslation} from '@atb/translations';
 import PaymentMethodsTexts from '@atb/translations/screens/subscreens/PaymentMethods';
 import {useFontScale} from '@atb/utils/use-font-scale';
-import queryString from 'query-string';
-import React, {useCallback, useEffect, useState} from 'react';
+import React from 'react';
 import {RefreshControl, View} from 'react-native';
 import {destructiveAlert} from './utils';
 import {FullScreenView} from '@atb/components/screen-view';
 import {ScreenHeading} from '@atb/components/heading';
-import {useListRecurringPaymentsQuery} from '@atb/ticketing/use-list-recurring-payments-query';
-import {useDeleteRecurringPaymentMutation} from '@atb/ticketing/use-delete-recurring-payment-mutation';
-import {useCancelRecurringPaymentMutation} from '@atb/ticketing/use-cancel-recurring-payment-mutation';
-import {APP_SCHEME} from '@env';
-import {
-  closeInAppBrowseriOS,
-  openInAppBrowser,
-} from '@atb/in-app-browser/in-app-browser';
-import {useAuthContext} from '@atb/auth';
-import firestore, {
-  FirebaseFirestoreTypes,
-} from '@react-native-firebase/firestore';
-import Bugsnag from '@bugsnag/react-native';
+
 import {PaymentBrand} from '@atb/modules/payment';
+import {useRecurringPayment} from '@atb/ticketing/use-recurring-payment';
 
 export const Profile_PaymentMethodsScreen = () => {
   const styles = useStyles();
   const {t} = useTranslation();
-
   const {
-    data: recurringPayment,
-    refetch: refetchRecurringPayment,
-    isError: recurringPaymentError,
-    isFetching: recurringPaymentLoading,
-  } = useListRecurringPaymentsQuery();
-
-  const {
-    mutateAsync: deleteRecurringPayment,
-    isError: deleteRecurringPaymentError,
-  } = useDeleteRecurringPaymentMutation();
-
-  const {
-    mutateAsync: cancelRecurringPayment,
-    isError: cancelRecurringPaymentError,
-  } = useCancelRecurringPaymentMutation();
-
-  const [awaitingRecurringPaymentId, setAwaitingRecurringPaymentId] =
-    useState<number>();
-
-  const callback = useCallback(async () => {
-    closeInAppBrowseriOS();
-    await refetchRecurringPayment();
-  }, [refetchRecurringPayment]);
-
-  // In cases where the recurring payment appears in firestore before the
-  // callback is called, we can cancel the payment flow and reload the recurring
-  // payment list.
-  useOnRecurringPaymentReceived({
-    recurringPaymentId: awaitingRecurringPaymentId,
-    callback,
-  });
-
-  const isError =
-    recurringPaymentError ||
-    deleteRecurringPaymentError ||
-    cancelRecurringPaymentError;
-
-  const addPaymentMethodCallbackHandler = async (url: string) => {
-    if (url.includes('response_code') && url.includes('recurring_payment_id')) {
-      const responseCode = queryString.parseUrl(url).query.response_code;
-      const paymentId = Number(
-        queryString.parseUrl(url).query.recurring_payment_id,
-      );
-      if (responseCode === 'OK') {
-        refetchRecurringPayment();
-      } else if (responseCode === 'Cancel') {
-        cancelRecurringPayment(paymentId);
-      }
-    }
-  };
-
-  const onAddRecurringPayment = async () => {
-    const callbackUrl = `${APP_SCHEME}://payment-method-callback`;
-    const response = await addPaymentMethod(callbackUrl);
-    setAwaitingRecurringPaymentId(response.data.recurring_payment_id);
-    await openInAppBrowser(
-      response.data.terminal_url,
-      'cancel',
-      callbackUrl,
-      addPaymentMethodCallbackHandler,
-    );
-  };
+    recurringPaymentLoading,
+    recurringPayment,
+    refetchRecurringPayment,
+    deleteRecurringPayment,
+    isError,
+    recurringPaymentError,
+    onAddRecurringPayment,
+  } = useRecurringPayment();
 
   return (
     <FullScreenView
@@ -251,52 +180,6 @@ const GenericError = () => {
       />
     </View>
   );
-};
-
-export const useOnRecurringPaymentReceived = ({
-  recurringPaymentId,
-  callback,
-}: {
-  recurringPaymentId: number | undefined;
-  callback: () => void;
-}) => {
-  const {userId} = useAuthContext();
-
-  const mapRecurringPaymentIds = (
-    d: FirebaseFirestoreTypes.DocumentSnapshot,
-  ): number => {
-    const recurringPayment = d.data();
-    if (!recurringPayment) {
-      throw new Error('No recurring payment data');
-    }
-
-    return recurringPayment.id;
-  };
-
-  useEffect(() => {
-    if (!recurringPaymentId) return;
-
-    const recurringPaymentsUnsub = firestore()
-      .collection('customers')
-      .doc(userId)
-      .collection('recurringPayments')
-      .onSnapshot(
-        (snapshot) => {
-          const recurringPaymentIds = snapshot.docs.map(mapRecurringPaymentIds);
-          if (recurringPaymentIds.some((id) => id === recurringPaymentId)) {
-            callback();
-          }
-        },
-        (err) => {
-          Bugsnag.notify(err, function (event) {
-            event.addMetadata('payment', {userId});
-          });
-        },
-      );
-    return () => {
-      recurringPaymentsUnsub();
-    };
-  }, [userId, recurringPaymentId, callback]);
 };
 
 const useStyles = StyleSheet.createThemeHook((theme: Theme) => ({
